traffic_model/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pillow
numpy
ultralytics
opencv-python